Finish implementation of thread file handling
This commit finishes the implementation of thread file handling, which
means that threads can now successfully be added or removed from the
data store. Duplicates are checked for and threads can only be added if
they exist within the server.

This commit also removes the need for channel IDs for threads. A thread
ID on its own works as a channel ID because threads are treated as
temporary channels.