Make small changes to command deployment
This commit removes an unnecessary usage of fs in the delete-commands.js
script, and it also accounts for weird behaviour where calling
deploy-commands.js will attempt to run App.js. I think this has to do
with deploy-commands.js requiring all the command files, and one of the
command files (RefreshThread) calls DiscordUtil, which calls App to
retrieve its instance to get the Discord client.

I'm not sure how to fix this behaviour but for now I'm putting a
band-aid on it.